refactor: remove notification demo module
